requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
scrapy>=2.11.0
playwright>=1.40.0

# Google Sheets Integration
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# WhatsApp Business API
heyoo>=0.1.0
twilio>=8.10.0

# Email
sendgrid>=6.10.0
python-dotenv>=1.0.0

# AI/GPT Integration
openai>=1.3.0

# Data Processing
pandas>=2.1.0
geopy>=2.4.0
python-dateutil>=2.8.0

# Utilities
pydantic>=2.5.0
python-json-logger>=2.0.7
schedule>=1.2.0

# PDF Generation
reportlab>=4.0.0
weasyprint>=60.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
